module restructure
I assume the fact that the enter button is whack, was because of some
blunder preceding this change

this seems like the module structure that would be best for unit tests,
even though each module contains multiple key
functions/responsibilities.

The important thing is each file is reasonably small, and there are no cycles

there are glob imports and public fields, the former might be ok, but
the latter should be changed somehow.